build.gradle: use slf4j-log4j12
Had it not been for Kafka I'd have intuitively gone with logback
and none of this would have happened.

That's my terrible excuse.

Luckily we've been driven more by metrics than log events, but now that
we see exceptions in the pipeline rather than the async kafka, we really
want to know what they are...